resolve GPDB_12_MERGE_FIXME in createplan.c
1. due to function push_partition_selector_candidate_for_join: for hashjoin,
or for nestloop or mergejoin with a Material or Sort node on the inner side,
they will process all the inner rows before scanning the outer side. Only in
these cases, a Partition Selector could be placed at this join. Even if the
prefetch_inner flag maybe reset to false, it doesn't affect the resullt.
2. the postgres planner doesn't support hash join on IS NOT DISTINCT FROM
for now, get rid of the hashqualclauses parameter from the make_hashjoin().