resolve GPDB_12_MERGE_FIXME in PartitionSelector to support for walking prune info
1.remove fixme in subselect.c: the paramid in PartitionSelector and the
params in part_prune_info is used to do partition pruning on the other
side of the join, they won't affect the output of this plan node. And
the params in part_prune_info are also in join qual or join filter,
so they won't affect the join plan's extParam/allParam either.
2.add the support to walk prune infos.