pg_dump: collect using sequence
When a table has an explicitly created sequence, the sequence is
ignored by pg_dump if it's not owned by the table.

Collect all sequences which maybe referenced by table columns. For
implicitly created sequence like SERIAL and GENERATED AS IDENTITY,
the sequence and the referenced table have dependency in pg_depend
with dependency type 'a' and 'i'. For explicitly created sequence
and SERIAL, they both have dependency with referenced attribute
definition. Consider both scenarios when identify used sequences.