Fix insert result error to sequence column
After rollback a transaction which execute restart identity clause
in, the insert result to sequence column is incorrect. It's due to
the cache value stored in hash table in QE has been out of date but
not be cleaned.

If the sequence has been transactionally replaced since we last saw
it, should discard any cached-but-unissued values both in QD and QE.
Dispatch TruncateStmt with restart_seqs to enable segment to reset a
sequence when specify restart identity.