fix unknown var coerce
An UNKNOWN var refered to the targetlist of a join's subquery in the
insert statement can't coerce to a proper type lead to error. When
coerce qury targetlist's tpe to the insert column's type, the fixed
unknown var can be coerce to other type which isn't equal to target
type. Do unknownout(Var) for a non-UNKNOWN is dangerous.

When transforms a Select Statement, Untyped Const or Param nodes in a
subquery in the FROM clause might have been assigned proper types when
we transformed the WHERE clause. So after that, when coerce UNKNOWN var
in targetlist to the assigned type, don't need to specify target type.
Now we simply pass UNKNOWNOID to the tartgettypeId, then if the var refs
to a join, it will still keep unchanged.

If the unknown var is refered to a join, the joinvar must refs to a leaf
RTE of the join, descend through leaf RTEs regardless of the targetType
will always bring targetlist Var types up to date. And when coerce qury
target's type to insert column, type of upper level keeps same as refer
vars.