[backport 6x] Unassign pure-catalog query from resource group
When gp_resource_group_bypass_catalog_query is on, unassign the pure
catalog query from resource group, and bypass the query.

Traverse the flattened rangetable entries to decide whether the query
is pure-catalog query.

In addition, it's hard to see whether the query contains UDF, traverse
the whole plan is expensive, so it will be ignored even if there can
have non-catalog query in the sql commands inside the UDF. But for the
special case like 'select UDF;', the query won't be bypassed. We have
no need to consider the sql commands inside the UDF, they will also be
bypassed or use the same resgroup as the outer query.